projet jeux olympiques final

+++ equipe.php
@@ -2,8 +2,10 @@
 
 include "_inc.php";
 
+$equipes = getEquipes();
+
 if( !empty($_POST['nom_equipe']) && empty($_POST['id_equipe']) ){
-     if( !equipeExists($_POST['nom_equipe']) ){
+     if( !equipeExists($_POST['nom_equipe']) && nameValid($_POST['nom_equipe']) ){
           $query = "INSERT INTO equipe VALUES(NULL, :nom)";
           $res = execReq($query, ["nom" => $_POST["nom_equipe"]]);
 
@@ -22,17 +24,18 @@ if( !empty($_POST['nom_equipe']) && empty($_POST['id_equipe']) ){
 }else if( isset($_GET['action']) && ctype_digit($_GET['id']) ){
      if( $_GET['action'] == "delete" ){
 
-          $res = execReq("DELETE FROM equipe WHERE id_equipe = :id", ['id' => $_GET['id']]);
+          if( !equipeExists($_GET['nom_equipe']) ){
+               $res = execReq("DELETE FROM equipe WHERE id_equipe = :id", ['id' => $_GET['id']]);
 
-          if( $res->rowCount() != 0 ){
-               $_SESSION['success'] = "Equipe supprimée avec success";
-          
-               header("location: equipe.php");
-               exit;
-          }else{
-               $_SESSION['warning'] = "delete equipe impossible";
+               if( $res->rowCount() != 0 ){
+                    $_SESSION['success'] = "Equipe supprimée avec success";
+               
+                    header("location: equipe.php");
+                    exit;
+               }
           }
-
+          
+          $_SESSION['warning'] = "delete equipe impossible";
 
 
      }else if( $_GET['action'] == "update" ){
@@ -54,9 +57,34 @@ if( !empty($_POST['nom_equipe']) && empty($_POST['id_equipe']) ){
 
           $equipeToUp = execReq("SELECT * FROM equipe WHERE id_equipe = :id", ['id' => $_GET['id']])->fetch(); 
      }
+}else if( isset($_POST['filtre']) ){
+     $filter = $_POST['filtre'];
+
+     if( $filter == "withMatch" ){
+          $query = "SELECT DISTINCT equipe.* 
+                    FROM equipe
+                    INNER JOIN rencontre
+                    ON equipe.id_equipe = rencontre.id_equipe_a
+                    OR equipe.id_equipe = rencontre.id_equipe_b
+                    WHERE date_rencontre >= now()";
+          $equipes = execReq($query)->fetchAll();
+     }else if( $filter == "withOutMatch" ){
+          $query = "SELECT DISTINCT equipe.* 
+                    FROM equipe
+                    WHERE id_equipe NOT IN
+                    (SELECT DISTINCT id_equipe_a 
+                    FROM rencontre)
+                    AND id_equipe NOT IN
+                    (SELECT DISTINCT id_equipe_b 
+                    FROM rencontre)";
+
+          $equipes = execReq($query)->fetchAll();
+     }
+
 }
 
-$equipes = getEquipes();
+
+
 
 include ("views/_header.php");
